Improve flexberry-acceptance-test blueprint
* Add tests with new test helpers
* Fix used test helpers
* Change default options for generated lists
* Add root directory and tag for all generated tests
* Add overwriting of generated tests by default
