--- blueprints/flexberry-acceptance-test/files/tests/acceptance/__name__-test.js
@@ -1,21 +1,0 @@
-import { test } from 'qunit';
-
-import moduleForAcceptance from '<%= testFolderRoot %>/tests/helpers/module-for-acceptance';
-
-moduleForAcceptance('<%= friendlyTestName %>');
-
-test('testing /<%= dasherizedModuleName %>', function(assert) {
-  visit('/<%= dasherizedModuleName %>');
-
-  andThen(() => {
-    checkOlvConfig('[data-test-olv]', null, assert, [
-      'refreshButton',
-      'createNewButton',
-      'colsConfigButton',
-    ]);
-
-    const controller = this.application.__container__.lookup('controller:' + currentRouteName());
-    const newFormRoute = controller.get('editFormRoute') + '.new';
-    goToNewForm('[data-test-olv]', null, assert, newFormRoute);
-  });
-});
