+++ blueprints/flexberry-list-form/files/app/templates/__name__.hbs
@@ -10,6 +10,7 @@
     refreshButton=true
     sorting=computedSorting
     orderable=true
+    showCheckBoxInRow=true
     sortByColumn=(action "sortByColumn")
     addColumnToSorting=(action "addColumnToSorting")
     beforeDeleteAllRecords=(action "beforeDeleteAllRecords")
