+++ tests/acceptance/components/flexberry-objectlistview/folv-sorting-on-all-columns-test.js
@@ -1,7 +1,6 @@
 import { executeTest } from './execute-folv-test';
 
 executeTest('check sorting on all column', (store, assert, app) => {
-  assert.expect(1);
   const path = 'components-acceptance-tests/flexberry-objectlistview/base-operations';
   visit(path);
   andThen(() => {
